Changed reordering animation
Instead of the bands moving with their labels, they are now static and only move when they have crossed the threshold of re-arranging.
This simplifies code and makes the visuals a bit easier to follow.
Calibration should be done to find optimum "crossover thresholds" for the labels, they seem a bit off.