Converted secondary controller to UINavController
Converted from UITableViewController to UINavigationController to allow for the pushing of new UITableViews when drilling down through constraint categories.
Also added UITableView as the sole view of the NavController, keeping all functionality intact (I hope).
To do so, had to fudge the 'initWithCoder:' call in the SecondaryViewController to call 'init' on super, rather than 'initWithCoder:', hopefully this doesn't call problems down the line.